# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "用于知夕番茄钟程序的后台api文档",
        "title": "知夕后台Api接口",
        "contact": {
            "name": "Doremy",
            "url": "https://dimdoremy.github.io/"
        },
        "license": {},
        "version": "1.0-golang"
    },
    "host": "127.0.0.1:8848",
    "basePath": "/api",
    "paths": {
...
        "/deleteuserdata": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-routers"
                ],
                "summary": "更新单个用户信息",
                "parameters": [
                    {
                        "description": "更新的用户信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    },
                    "400": {
                        "description": "400 bad request",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        },
...
        "/insertuserdata": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-routers"
                ],
                "summary": "注册单个用户",
                "parameters": [
                    {
                        "description": "注册用户信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    },
                    "400": {
                        "description": "注册失败",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        },
...
        "/querybyopenid": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-routers"
                ],
                "summary": "获取单个用户信息",
                "parameters": [
                    {
                        "description": "单个用户信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "单个用户信息的json数据",
                        "schema": {
                            "$ref": "#/definitions/model.UserData"
                        }
                    },
                    "400": {
                        "description": "400 bad request",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        },
...
        "/updateuserdata": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-routers"
                ],
                "summary": "更新单个用户信息",
                "parameters": [
                    {
                        "description": "更新的用户信息",
                        "name": "handBook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    },
                    "400": {
                        "description": "400 bad request",
                        "schema": {
                            "$ref": "#/definitions/model.MessageBind"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.MessageBind": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.UserData": {
            "type": "object",
            "properties": {
                "avatarurl": {
                    "type": "string"
                },
                "experience": {
                    "type": "integer"
                },
                "lifepoint": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "openid": {
                    "type": "string"
                },
                "soul": {
                    "type": "integer"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
...
    }
}